Implementation and testing of linear Lagrangian finite elements
