+++ examples/assemble/Dirichlet_Laplacian_demo.cc
@@ -17,278 +17,9 @@
 #include <lf/io/io.h>
 #include <lf/mesh/hybrid2dp/hybrid2dp.h>
 #include <lf/refinement/refinement.h>
+#include "lf/fe/fe.h"
 #include "lf/mesh/test_utils/test_meshes.h"
 #include "lf/mesh/utils/utils.h"
-
-/** @brief Local assembler model class for linear Lagrangian finite elements
- *
- * The main purpose of this class is to compute the element matrix for
- * the Laplacian on affine triangles or bilinearly mapped quadrilaterals.
- * These element matrices are provided by the `Eval()` method.
- *
- * @note the `Eval()` method will always return a _reference_ to a 4x4 matrix
- * also for triangles. In this case the last row and column must be ignored.
- */
-class LinearFELaplaceElementMatrix {
- public:
-  using elem_mat_t = Eigen::Matrix<double, 4, 4>;
-  using ElemMat = const elem_mat_t;
-
-  LinearFELaplaceElementMatrix(const lf::mesh::Mesh &mesh) : mesh_(mesh) {}
-  bool isActive(const lf::mesh::Entity &cell) { return true; }
-  ElemMat Eval(const lf::mesh::Entity &cell);
-
- private:
-  /// Reference to current mesh
-  const lf::mesh::Mesh &mesh_;
-  /// quadrature points on reference quadrilateral
-  const double kSqrt3 = 1.0 / std::sqrt(3.0);
-  const double kZeta_0 = 0.5 - 0.5 * kSqrt3;
-  const double kZeta_1 = 0.5 + 0.5 * kSqrt3;
-  const std::array<Eigen::Vector2d, 4> kQuadPoints{
-      Eigen::Vector2d{kZeta_0, kZeta_0}, Eigen::Vector2d{kZeta_0, kZeta_1},
-      Eigen::Vector2d{kZeta_1, kZeta_0}, Eigen::Vector2d{kZeta_1, kZeta_1}};
-  // Gradients of refrence shape functions in rows of a matrix
-  static Eigen::Matrix<double, 4, 2> DervRefShapFncts(
-      const Eigen::Vector2d &xh);
-  // Barycenter of triangle
-  const Eigen::Matrix<double, 2, 1> kTriabc{
-      (Eigen::Matrix<double, 2, 1>() << (1.0 / 3), (1.0 / 3)).finished()};
-  // Point in reference square
-  const Eigen::Matrix<double, 2, 1> kQuadpt{
-      (Eigen::Matrix<double, 2, 1>() << 0.7, 0.3).finished()};
-
- public:
-  static unsigned int dbg_ctrl;
-  static const unsigned int dbg_det = 1;
-  static const unsigned int dbg_locmat = 2;
-  static const unsigned int dbg_J = 4;
-  static const unsigned int dbg_geo = 8;
-};
-
-unsigned int LinearFELaplaceElementMatrix::dbg_ctrl{0};
-
-inline Eigen::Matrix<double, 4, 2>
-LinearFELaplaceElementMatrix::DervRefShapFncts(const Eigen::Vector2d &xh) {
-  // clang-format off
-  return (Eigen::Matrix<double, 4, 2>(4, 2) <<
-	  xh[1] - 1, xh[0] - 1,
-	  1 - xh[1],  -xh[0]  ,
-	  xh[1]    , xh[0]    ,
-	  -xh[1]   , 1 - xh[0]
-	  ).finished();
-  // clang-format on
-}
-
-LinearFELaplaceElementMatrix::ElemMat LinearFELaplaceElementMatrix::Eval(
-    const lf::mesh::Entity &cell) {
-  // Topological type of the cell
-  const lf::base::RefEl ref_el{cell.RefEl()};
-
-  // Obtain the vertex coordinates of the cell, which completely
-  // describe its shape.
-  const lf::geometry::Geometry *geo_ptr = cell.Geometry();
-  LF_ASSERT_MSG(geo_ptr != nullptr, "Invalid geometry!");
-  LF_ASSERT_MSG((geo_ptr->DimGlobal() == 2) && (geo_ptr->DimLocal() == 2),
-                "Only 2D implementation available!");
-  const Eigen::MatrixXd &ref_el_corners(ref_el.NodeCoords());
-  // Matrix storing corner coordinates in its columns
-  const Eigen::MatrixXd vertices{geo_ptr->Global(ref_el_corners)};
-  // Debugging output
-  SWITCHEDSTATEMENT(dbg_ctrl, dbg_geo,
-                    std::cout << ref_el << '(' << mesh_.Index(cell)
-                              << "), shape = \n"
-                              << vertices << std::endl);
-
-  // 4x4 dense matrix for returning result
-  elem_mat_t elem_mat = elem_mat_t::Zero(4, 4);
-
-  // Computations differ depending on the type of the cell
-  switch (ref_el) {
-    case lf::base::RefEl::kTria(): {
-      LF_ASSERT_MSG((vertices.cols() == 3) && (vertices.rows() == 2),
-                    "Wrong size of vertex matrix!");
-      // Obtain area (not needed, can also be obtained from determinant
-      // of auxiliary matrix
-      // Eigen::Matrix<double, 2, 1> refc;
-      // refc << 1.0 / 3, 1.0 / 3;
-      // const double area = 0.5 * (geo_ptr->IntegrationElement(refc))[0];
-
-      // Set up an auxiliary 3x3-matrix with a leading column 1 and
-      // the vertex coordinates in its right 3x2 block
-      Eigen::Matrix<double, 3, 3> X;  // temporary matrix
-      X.block<3, 1>(0, 0) = Eigen::Vector3d::Ones();
-      X.block<3, 2>(0, 1) = vertices.transpose();
-      // The determinant of the auxliriary matrix also supplies the determinant
-      const double area = 0.5 * std::abs(X.determinant());
-      SWITCHEDSTATEMENT(dbg_ctrl, dbg_det,
-                        std::cout
-                            << "Area: " << area << " <-> "
-                            << 0.5 * (geo_ptr->IntegrationElement(kTriabc))
-                            << std::endl);
-
-      // Compute the gradients of the barycentric coordinate functions
-      // and store them in the columns of a 2x3 matrix grad_bary_coords
-      Eigen::Matrix<double, 2, 3> grad_bary_coords{
-          X.inverse().block<2, 3>(1, 0)};
-      // Since the gradients are constant local integration is easy
-      elem_mat.block<3, 3>(0, 0) =
-          area * grad_bary_coords.transpose() * grad_bary_coords;
-      break;
-    }
-    case lf::base::RefEl::kQuad(): {
-      LF_ASSERT_MSG((vertices.cols() == 4) && (vertices.rows() == 2),
-                    "Wrong size of vertex matrix!");
-      // Coefficient matrix for bilinear mapping to actual cell
-      // clang-format off
-      Eigen::Matrix<double, 2, 4> G(vertices *
-                                    (Eigen::Matrix<double, 4, 4>() <<
-				     1, -1, -1, 1,
-				     0, 1 , 0, -1,
-				     0, 0, 0, 1,
-				     0, 0, 1, -1)
-                                        .finished());
-      // clang-format on 
-      // Coefficients for the determinant as a linear function in
-      // the reference coordinates
-      // clang-format off
-      Eigen::Vector3d detc{
-          (Eigen::Vector3d() <<
-	   G(0, 1) * G(1, 2) - G(1, 1) * G(0, 2),
-           G(0, 1) * G(1, 3) - G(1, 1) * G(0, 3),
-           G(1, 2) * G(0, 3) - G(0, 2) * G(1, 3))
-              .finished()};
-      // clang-format on
-      // Determinant at a point in the reference element
-      auto detDPhi = [&detc](const Eigen::Vector2d &xh) -> double {
-        return std::abs(detc[0] + detc[1] * xh[0] + detc[2] * xh[1]);
-      };
-      SWITCHEDSTATEMENT(
-          dbg_ctrl, dbg_det,
-          std::cout << "Determinant: " << detDPhi(kQuadpt) << " <-> "
-                    << (geo_ptr->IntegrationElement(kQuadpt)) << std::endl);
-
-      // Transposed adjunct matrix of Jacobian of transformation
-      auto DPhiadj =
-          [&G](const Eigen::Vector2d &xh) -> Eigen::Matrix<double, 2, 2> {
-        // clang-format off
-        return ((Eigen::Matrix<double, 2, 2>() <<
-		 G(1, 2) + G(1, 3) * xh[0] , -G(1, 1) - G(1, 3) * xh[1],
-		 -G(0, 2) - G(0, 3) * xh[0], G(0, 1) + G(0, 3) * xh[1])
-                    .finished());
-        // clang-format on
-      };
-      // Output for debugging
-      SWITCHEDSTATEMENT(
-          dbg_ctrl, dbg_J,
-          std::cout << "InverseTransposedJacobian:\n"
-                    << (DPhiadj(kQuadpt) / detDPhi(kQuadpt)) << " <-> "
-                    << (geo_ptr->JacobianInverseGramian(kQuadpt)) << std::endl);
-
-      // For a quadrilateral we have to use numerical quadrature based on
-      // a 2-2 tensor-product Gauss rule.
-      // Sum over quadrature points (weight= 0.5)
-      for (int q = 0; q < 4; ++q) {
-        // Obtain reference coordinates of current quadrature point
-        const Eigen::Vector2d qp(kQuadPoints[q]);
-        // Gradients of shape functions in quadrature points stored in
-        // the columns of a matrix
-        Eigen::Matrix<double, 2, 4> gradmat(DPhiadj(qp) *
-                                            DervRefShapFncts(qp).transpose());
-        // Update of element matrix by contribution from current quadrature
-        // point, whose entries are dot products of gradients
-        elem_mat += gradmat.transpose() * gradmat / detDPhi(qp);
-      }
-      // Scaling with quadrature weights
-      elem_mat *= 0.25;
-      break;
-    }
-    default: { LF_ASSERT_MSG(false, "Illegal cell type"); }
-  }  // end switch
-
-  SWITCHEDSTATEMENT(
-      dbg_ctrl, dbg_locmat, lf::base::size_type nv = vertices.cols();
-      std::cout << "Element matrix\n"
-                << elem_mat.block(0, 0, nv, nv) << std::endl;
-      std::cout << "Row sums = " << elem_mat.block(0, 0, nv, nv).colwise().sum()
-                << ",\n col sums = "
-                << elem_mat.block(0, 0, nv, nv).rowwise().sum().transpose()
-                << std::endl);
-
-  // Return the element matrix
-  return elem_mat;
-}
-
-/** @brief Class for computation of local load vector for linear finite
- * elements.
- *
- * @tparam FUNCTOR object with an evaluation operator of signature
- *         std::function<double(const Eigen::Vector2d &)>, which supplies
- *         the source function
- *
- * Computation is based on vertex based quadrature
- *
- * @note The element vector returned by the `Eval()` method will always
- * have length 4 also for triangles
- */
-template <typename FUNCTOR>
-class LinearFELocalLoadVector {
- public:
-  using elem_vec_t = Eigen::Matrix<double, 4, 1>;
-  using ElemVec = const elem_vec_t;
-
-  LinearFELocalLoadVector(const lf::mesh::Mesh &mesh, FUNCTOR f)
-      : mesh_(mesh), f_(f) {}
-  bool isActive(const lf::mesh::Entity &cell) { return true; }
-  ElemVec Eval(const lf::mesh::Entity &cell);
-
- private:
-  const lf::mesh::Mesh &mesh_;
-  FUNCTOR f_;
-
- public:
-  static unsigned int dbg_ctrl;
-  static const unsigned int dbg_locvec = 1;
-  static const unsigned int dbg_geo = 2;
-};
-
-template <typename FUNCTOR>
-unsigned int LinearFELocalLoadVector<FUNCTOR>::dbg_ctrl = 0;
-
-template <typename FUNCTOR>
-typename LinearFELocalLoadVector<FUNCTOR>::ElemVec
-LinearFELocalLoadVector<FUNCTOR>::Eval(const lf::mesh::Entity &cell) {
-  // Topological type of the cell
-  const lf::base::RefEl ref_el{cell.RefEl()};
-  const lf::base::size_type num_nodes{ref_el.NumNodes()};
-
-  // Obtain the vertex coordinates of the cell, which completely
-  // describe its shape.
-  const lf::geometry::Geometry *geo_ptr = cell.Geometry();
-  LF_ASSERT_MSG(geo_ptr != nullptr, "Invalid geometry!");
-  LF_ASSERT_MSG((geo_ptr->DimGlobal() == 2) && (geo_ptr->DimLocal() == 2),
-                "Only 2D implementation available!");
-  const Eigen::MatrixXd &ref_el_corners(ref_el.NodeCoords());
-  const Eigen::MatrixXd vertices{geo_ptr->Global(ref_el_corners)};
-  // Debugging output
-  SWITCHEDSTATEMENT(dbg_ctrl, dbg_geo,
-                    std::cout << ref_el << ", shape = \n"
-                              << vertices << std::endl);
-  const double area = lf::geometry::Volume(*geo_ptr);
-
-  // Vector for returning element vector
-
-  elem_vec_t elem_vec = elem_vec_t::Zero();
-  // get function values in the vertices
-  for (int k = 0; k < num_nodes; k++) {
-    elem_vec[k] = area * f_(vertices.col(k)) / num_nodes;
-  }
-  SWITCHEDSTATEMENT(
-      dbg_ctrl, dbg_locvec,
-      std::cout << "element vector = " << elem_vec.head(num_nodes).transpose()
-                << std::endl);
-  return elem_vec;
-}
 
 static unsigned int dbg_ctrl = 0;
 const unsigned int dbg_dofh = 1;
@@ -419,8 +150,8 @@ double L2ErrorLinearFEDirichletLaplacian(
       lf::mesh::utils::PrintInfo(*mesh_p, std::cout);
       lf::mesh::hybrid2dp::Mesh::output_ctrl_ = tmp_mesh_ctrl);
   // Initialize objects for local computations
-  LinearFELaplaceElementMatrix loc_mat_laplace(*mesh_p);
-  LinearFELocalLoadVector loc_vec_sample(*mesh_p, f);
+  lf::fe::LinearFELaplaceElementMatrix loc_mat_laplace(*mesh_p);
+  lf::fe::LinearFELocalLoadVector loc_vec_sample(*mesh_p, f);
   // Initialization of index mapping for linear finite elements
   lf::assemble::UniformFEDofHandler loc_glob_map(
       mesh_p, {{lf::base::RefEl::kPoint(), 1}});
@@ -630,10 +361,10 @@ int main(int argc, const char **argv) {
     }
 
     // Set debugging switches
-    LinearFELaplaceElementMatrix::dbg_ctrl = 0;
+    lf::fe::LinearFELaplaceElementMatrix::dbg_ctrl = 0;
     // LinearFELaplaceElementMatrix::dbg_geo |
     // LinearFELaplaceElementMatrix::dbg_locmat;
-    LinearFELocalLoadVector<decltype(f)>::dbg_ctrl = 0;
+    lf::fe::LinearFELocalLoadVector<decltype(f)>::dbg_ctrl = 0;
     lf::assemble::DofHandler::output_ctrl_ = 6;
     dbg_ctrl = dbg_basic;  // | dbg_mat | dbg_mesh | dbg_dofh | dbg_trp;
     // lf::assemble::ass_mat_dbg_ctrl = 255;
