+++ lib/lf/fe/CMakeLists.txt
@@ -1,6 +1,9 @@
 set(sources
   lagr_fe.h
   lagr_fe.cc
+  loc_comp_ellbvp.h
+  loc_comp_ellbvp.cc
+  fe.h
 )
 
 add_library(lf.fe ${sources})
@@ -9,4 +12,4 @@ target_link_libraries(lf.fe PUBLIC
   lf.mesh.utils lf.assemble)
 target_compile_features(lf.fe PUBLIC cxx_std_17)
 
-# add_subdirectory(test)
+add_subdirectory(test)
